requirements.txt
aiogram
openai
openai-whisper
pydantic
pydantic-settings
python-dotenv
